Static Site Changes
I changed the jumbotron bootstrap code and edit the About, Contact and I added a gallery page.

The gallery page has a simple gallery that we can change in the future to include what you have created of course.

Please revise all of the code changes that I have made and I hope that I am doing this correctly.